Syslog supports JSON.
syslog-ng now can receive messages either with or without JSON. It
always sends JSON out to Kafka. If the original message was JSON, it
will be parsed and included in the "_parsed" field of the resulting JSON
output.

For now this is still just stored raw in the MySQL database, since it's
not worth putting in extra effort into that DB at this point.