Determine the origin of a message based on what it looks like.
This assumes that it receives parsed JSON from syslog-ng, and then uses
core.match to pattern match its structure.

We've also tied this into the Onyx catalog, and shifted around the BS
MySQL stuff so we store the origin in the table. (This is done stupidly,
but we're just going to throw it all away when we drop MySQL anyway.)