Make Docker ignore node_modules
Otherwise, Docker will copy this from your development environment. If
that's OS X, that will include fsevents, a library that only works on OS
X. That wouldn't be _too_ bad, except that expects the node binary to be
called "node" (as it is in upstream node), which isn't true on Debian,
where it is called "nodejs".